requirements.txt
Django==2.2.7
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
npm==0.1.1
optional-django==0.1.0
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0

